requirements.txt
pyTelegramBotAPI>=4.20,<5.0
python-dotenv>=1.0,<2.0
requests>=2.32,<3.0
psycopg2-binary>=2.9,<3.0
pytest>=8.0,<9.0
